check in defects

+++ WebERP/Models/CuttingReceipt/Cutting_Receipt.cs
@@ -14,7 +14,7 @@ namespace WebERP.Models
         public int DOC_FINYEAR { get; set; }
         public int DOC_NO { get; set; }
         public int CUTTING_ORDER_FK { get; set; }
-        public decimal RECEIPT_QTY { get; set; }
+        public int RECEIPT_QTY { get; set; }
         public string EMP_NAME { get; set; }
         public string ITEM_NAME { get; set; }
         public string ART_NAME { get; set; }
